Abstracting code after teacher review
