+++ src/js/render/posts/commentForm.mjs
@@ -1,0 +1,9 @@
+import { createComment } from "../../templates/index.mjs";
+
+export function displayCommentForm() {
+  const createCommentForm = document.querySelector("#createCommentForm");
+
+  if (createCommentForm) {
+    createCommentForm.innerHTML = createComment();
+  }
+}
