+++ src/js/render/posts/postOptions.mjs
@@ -1,0 +1,15 @@
+import { get } from "../../storage/index.mjs";
+import { postOptions } from "../../templates/index.mjs";
+
+export function displayPostOptions(author) {
+  const dropdown = document.querySelector("#dropdownContainer");
+
+  if (dropdown) {
+    dropdown.innerHTML = "";
+    const loggedInUser = get("user").name;
+
+    if (loggedInUser === author) {
+      dropdown.innerHTML = postOptions();
+    }
+  }
+}
