+++ src/js/render/posts/renderPost.mjs
@@ -1,7 +1,8 @@
 import * as postMethods from "../../api/posts/index.mjs";
 import * as templates from "../../templates/index.mjs";
+import { displayCommentForm } from "./commentForm.mjs";
+import { displayPostOptions } from "./postOptions.mjs";
 import { noResultError } from "../../components/error.mjs";
-import { get } from "../../storage/index.mjs";
 
 /**
  * Displaying a single post details on the page using API call
@@ -17,24 +18,8 @@ export async function renderPostHtml() {
       postContainer.innerHTML = templates.postTemplate(result);
     }
 
-    // if post contain no image, hide img tag from html template
-    const media = document.querySelector(".media");
-
-    if (result.media === "") {
-      media.style.display = "none";
-    }
-
     // buttons to edit or delete post if the post belongs to the logged in user
-    const dropdown = document.querySelector("#dropdownContainer");
-
-    if (dropdown) {
-      dropdown.innerHTML = "";
-      const loggedInUser = get("user").name;
-
-      if (loggedInUser === result.author.name) {
-        dropdown.innerHTML = templates.postOptions();
-      }
-    }
+    displayPostOptions(result.author.name);
 
     // displaying comments if there is any
     const commentSection = document.querySelector(".comment-section");
@@ -53,9 +38,5 @@ export async function renderPostHtml() {
   }
 
   // displaying comment form regardless if there already are comments or not
-  const createCommentForm = document.querySelector("#createCommentForm");
-
-  if (createCommentForm) {
-    createCommentForm.innerHTML = templates.createComment();
-  }
+  displayCommentForm();
 }
